Allow the map to be zoomed using the mousewheel
Previously, using the mousewheel would always send synthetic up and down
keys, which would cause units to be ordered up or down when they were
selected on the map screen, which is unexpected.

The edge and drag scrolling code doesn't work without the smooth
scrolling patches, but before this patch some of their code would still
run anyway when mouse events are received by the window procedure. The
scrolling code relied on map_is_visible() always returning false when
smooth scrolling was disabled so that it would return early
(map_is_visible() would return false because it checked if
CState.TimersEnabled == true, but it is only set to true in a patch that
is only installed when smooth scrolling is enabled).

This patch changes the window procedure to skip all mouse handling when
smooth scrolling is disabled so that clicking and dragging the right
mouse does not cause strange behaviour. This also makes the code
clearer.